chore: Use secrets for sensitive data
Prefer compose element 'secrets' for sensitive data over env variables.
Add 'db' dir in project root to host DB data and files.

[Note]
Compose raises an error if a secret file is not added.
Add example file: '_db_password.txt' for user to rename and set secret.
Add double quotes ("") which triggers an error if user tries to use it
as a secret passwd for the Postgres container.