chore: Add pre-commit hook (shellcheck)
Add hook `shellcheck-py`: Linter for shell scripts.
Use the `-py` wrapper version of the `shellcheck` binary.

[Note]
Prefer the Python wrapper version due to the following:
- Installs prebuilt binaries of ShellCheck into your environment.
- Decouples dependency on system pkg managers or manual installations.
- Avoids system installation in `CI` pipelines where  can be cumbersome.

[Warning]
`shellcheck-py` is manually updated and synced with it's source binary.
This warrants users to verify that it is up-to-date with source version.

[Docs]
See, ShellCheck:
https://github.com/koalaman/shellcheck
See recommended ShellCheck pre-commit hook:
https://github.com/scop/pre-commit-shfmt
See (used) Python wrapper alternative:
https://github.com/shellcheck-py/shellcheck-py