chore: Add pre-commit hook (shfmt)
Add hook `shfmt-py`: Formatter for shell scripts.
Use the `-py` wrapper version of the `shfmt` binary.

[Note]
Prefer the Python wrapper version due to the following:
- Installs prebuilt binaries of `shfmt` into your environment.
- Decouples dependency on system pkg managers or manual installations.
- Avoids system installation in `CI` pipelines where  can be cumbersome.

[Warning]
`shfmt-py` is manually updated and synced with it's source binary.
This warrants users to verify that it is up-to-date with source version.

[Docs]
See, shfmt:
https://github.com/mvdan/sh
See recommended shfmt pre-commit hook:
https://github.com/scop/pre-commit-shfmt
See (used) Python wrapper alternative:
https://github.com/MaxWinterstein/shfmt-py