revert: Migrate secrets back to type `file`
Use secrets of type `file` for local dev and potentially prod.

[Reason]
External secrets are provided at runtime by the Docker platform. This
requires Docker running in swarm mode and the app stack deployed
on a manger node.
This cluster service management is overhead for local
development and does not justify the using external secrets.

[Commit]
- Name: refactor: Migrate secrets to type `external`
- SHA: 60c6abe6449cd9f2c0652d54829e8a2a8e8a21c7

[Note]
- A revert was required due to pushing the commit